+++ src/App.tsx
@@ -1,29 +1,14 @@
 import { Component } from "react";
 
 import Header from "./components/Header";
-import RestaurantFilterContainer from "./components/RestaurantFilterContainer";
-import RestaurantList from "./components/RestaurantList";
+import Main from "./components/Main";
 
-interface State {
-  category: string;
-  sorting: string;
-}
-
-class App extends Component<any, State> {
-  state = { category: "전체", sorting: "이름순" };
-
-  setOption(key: string, option: string): void {
-    this.setState((prev) => ({ ...prev, [key]: option }));
-  }
-
+class App extends Component {
   render() {
     return (
       <>
         <Header />
-        <main className="container">
-          <RestaurantFilterContainer setOption={this.setOption.bind(this)} />
-          <RestaurantList options={this.state} />
-        </main>
+        <Main />
       </>
     );
   }
