+++ src/components/Main.module.css
@@ -1,0 +1,4 @@
+.container {
+  max-width: 840px;
+  margin: 0 auto;
+}
