+++ src/styles/index.css
@@ -24,7 +24,4 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
 }
 
-.container {
-  max-width: 840px;
-  margin: 0 auto;
-}
+
